# openapi-BuyBuy.yaml
openapi: 3.0.3
info:
  title: BuyBuy
  version: 1.0.0
  description: An API for an e-commerce web application
servers:
  - url: http://localhost:4001
    variables: {}
    description: Local development server
paths:
  /users:
    summary: Access all users
    description: This path will grant access to all the users from the postgres database
    get:
      summary: Gets the users' data
      description: Gets the users' data from the postgres database
      operationId: get_users
      responses:
        '200':
          content:
            application/json:
              examples:
                users:
                  value: |-
                    {"users": [
                      {"id": "0000000001", "name": "Rachel Chan", 
                      "username": "rachelc1", "email": "rachelc@email.com", 
                      "phone": "+60110000000", "address_id": "0000000001", "seller": true}, 
                      {"id": "0000000002", "name": "Michael Chan", 
                      "username": "michaelc1", "email": "michaelc@email.com", "phone": "+60110000001", "address_id": "0000000002", "seller": false}, 
                      {"id": "0000000003", "name": "Amber Chan", 
                      "username": "amberc1", "email": "amberc@email.com", "phone": "+60110000002", "address_id": null, "seller": true}
                    ]}
          description: Success
        '400':
          content:
            text/plain; charset=utf-8:
              examples:
                error:
                  value: Bad request
          description: Bad request
      tags:
        - Users
    post:
      summary: Register a new user
      description: This operation adds a new user to the postgres database
      operationId: post_user
      requestBody:
        description: A new user object
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '201':
          content:
            application/json:
              examples:
                newUser:
                  value: |-
                    {"message": "User created successfully",
                      "userId": "0000000003"}
          description: Success
        '400':
          content:
            text/plain; charset=utf-8:
              examples:
                error:
                  value: Bad request
          description: Bad request
      tags:
        - Users
  /users/{userId}:
    summary: Access a specific user
    description: >-
      This path will grant access to the user with a matching ID from the 
      postgres database
    get:
      summary: Gets a specific user's data
      description: >-
        This operation will get the user's data with a matching ID from the 
        postgres database
      operationId: get_user
      parameters:
        - $ref: '#/components/parameters/userId'
      responses:
        '200':
          content:
            application/json:
              examples:
                users:
                  value: |-
                    {"id": "0000000002", "name": "Amber Chan", 
                      "email": "amberc@email.com", "phone": "+60110000002", 
                      "address_id": "0000000001", "seller": false}
          description: Success
        '400':
          content:
            text/plain; charset=utf-8:
              examples:
                error:
                  value: Bad request
          description: Bad request
      tags:
        - Users
    put:
      summary: Updates a specific user's data
      description: >-
        This operation will update the user's data with a matching ID from the 
        postgres database
      operationId: put_user
      parameters:
        - $ref: '#/components/parameters/userId'
      requestBody:
        description: An updated user object
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '200':
          content:
            application/json:
              examples:
                users:
                  value: |-
                    {"id": "0000000002", "name": "Amber Chan", 
                      "email": "amberc@email.com", "phone": "+60120000002", 
                      "address_id": "0000000002", "seller": true}
          description: Success
        '400':
          content:
            text/plain; charset=utf-8:
              examples:
                error:
                  value: Bad request
          description: Bad request
      tags:
        - Users
    delete:
      summary: Deletes a specific user's data
      description: >-
        This operation will delete the user's data with a matching ID from the 
        postgres database
      operationId: delete_user
      parameters:
        - $ref: '#/components/parameters/userId'
      responses:
        '204':
          description: Success
        '400':
          content:
            text/plain; charset=utf-8:
              examples:
                error:
                  value: Bad request
          description: Bad request
      tags:
        - Users
  /addresses:
    summary: Access all addresses
    description: This path will grant access to all the addresses from the postgres database
    get:
      summary: Gets the addresses data
      description: Gets the addresses data from the postgres database
      operationId: get_addresses
      responses:
        '200':
          content:
            application/json:
              examples:
                users:
                  value: |-
                    {"addresses": [
                      {"id": "0000000001", "unit_number": "1", 
                      "floor_number": "1", "building_name": "Imazium", 
                      "street": "Baker Street", "city": "Metropolis",
                      "region_id": 3, "zone": "Johor", "country": "Malaysia"}, 
                      {"id": "0000000002", "unit_number": "3", 
                      "floor_number": "1", "building_name": "Imazium", 
                      "street": "Baker Street", "city": "Metropolis",
                      "region_id": 3, "zone": "Johor", "country": "Malaysia"}, 
                      {"id": "0000000003", "unit_number": "3", 
                      "floor_number": "1", "building_name": "Imazium", 
                      "street": "Baker Street", "city": "Metropolis",
                      "region_id": 3, "zone": "Johor", "country": "Malaysia"}
                    ]}
          description: Success
        '400':
          content:
            text/plain; charset=utf-8:
              examples:
                error:
                  value: Bad request
          description: Bad request
      tags:
        - Addresses
    post:
      summary: Register a new address
      description: This operation adds a new address to the postgres database
      operationId: post_address
      requestBody:
        description: A new address object
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Address'
      responses:
        '201':
          content:
            application/json:
              examples:
                newUser:
                  value: |-
                    {
                      "message": "New address created", 
                      "addressId": "0000000001"
                    }
          description: Success
        '400':
          content:
            text/plain; charset=utf-8:
              examples:
                error:
                  value: Bad request
          description: Bad request
      tags:
        - Addresses
  /addresses/{addressId}:
    summary: Access all addresses
    description: This path will grant access to all the addresses from the postgres database
    get:
      summary: Gets the specific address data
      description: Gets the address data with the matching addressId from the postgres database
      operationId: get_address
      parameters:
        - $ref: '#/components/parameters/addressId'
      responses:
        '200':
          content:
            application/json:
              examples:
                users:
                  value: |-
                    {"id": "0000000001", "unit_number": "1", 
                      "floor_number": "1", "building_name": "Imazium", 
                      "street": "Baker Street", "city": "Metropolis",
                      "region_id": 3, "zone": "Johor", "country": "Malaysia"}
          description: Success
        '400':
          content:
            text/plain; charset=utf-8:
              examples:
                error:
                  value: Bad request
          description: Bad request
      tags:
        - Addresses
    put:
      summary: Updates the specific address data
      description: Updates the address data with the matching addressId from the postgres database
      operationId: put_address
      parameters:
        - $ref: '#/components/parameters/addressId'
      requestBody:
        description: A new address object
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Address'
      responses:
        '200':
          content:
            application/json:
              examples:
                address:
                  value: |-
                    {"id": "0000000001", "unit_number": "1", 
                      "floor_number": "1", "building_name": "Imazium", 
                      "street": "Baker Street", "city": "Metropolis",
                      "region_id": 3, "zone": "Johor", "country": "Malaysia"}
          description: Success
        '400':
          content:
            text/plain; charset=utf-8:
              examples:
                error:
                  value: Bad request
          description: Bad request
      tags:
        - Addresses
    delete:
      summary: Deletes the specific address data
      description: Deletes the address data with the matching addressId from the postgres database
      operationId: delete_address
      parameters:
        - $ref: '#/components/parameters/addressId'
      responses:
        '204':
          description: Success
        '400':
          content:
            text/plain; charset=utf-8:
              examples:
                error:
                  value: Bad request
          description: Bad request
      tags:
        - Addresses
...
components:
  schemas:
    User:
      type: object
      properties:
        name:
          type: string
        username:
          type: string
        email:
          type: string
        password:
          type: string
        phone:
          type: string
        seller:
          type: boolean
        address_id:
          type: string
    Address:
      type: array
      items:
        type: object
        properties:
          unitNumber:
            type: string
          floorNumber:
            type: integer
          street:
            type: string
          city:
            type: string
          postcode:
            type: string
          zone:
            type: string
          country:
            type: string
...
  parameters:
    userId:
      name: userId
      in: path
      description: The id of the user
      required: true
      schema:
        type: string
    addressId:
      name: addressId
      in: path
      description: The id of the address
      required: true
      schema:
        type: string
...